pyproject.toml
[project]
name = "weather-chatwork-notifier-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.36",
    "google-adk>=1.14.1",
    "litellm>=1.77.3",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]
